requirements.txt
asgiref==3.6.0
autopep8==2.0.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.5
django-cron==0.6.0
django-dotenv==1.4.2
django-ipware==4.0.2
idna==3.4
mysqlclient==2.1.1
pkg_resources==0.0.0
pycodestyle==2.10.0
requests==2.28.1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13